+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 ## UNRELEASED
 
 ### Bug Fixes
+
 - [cronos#109](https://github.com/crypto-org-chain/cronos/issues/109) ibc transfer timeout too short
+- [tharsis#590](https://github.com/tharsis/ethermint/pull/590) fix export contract state in genesis and reimport
 
 ### Features